Implemented new compiler which uses unitaries generated within components to vastly simplify compiler logic
