+++ lightworks/sdk/circuit/compiler.py
@@ -1,0 +1,124 @@
+# Copyright 2024 Aegiq Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from copy import copy
+
+import numpy as np
+
+from ..utils import ModeRangeError
+from .components import Component, Loss
+
+
+class CompiledCircuit:
+    """
+    New Compiler
+    """
+
+    def __init__(self, n_modes: int) -> None:
+        self._n_modes = n_modes
+        self._loss_modes = 0
+        self._unitary = np.identity(n_modes)
+        self._in_heralds: dict[int, int] = {}
+        self._out_heralds: dict[int, int] = {}
+
+    @property
+    def n_modes(self) -> int:
+        """Number of real modes."""
+        return self._n_modes
+
+    @property
+    def loss_modes(self) -> int:
+        """Number of loss modes."""
+        return self._loss_modes
+
+    @property
+    def total_modes(self) -> int:
+        """Total number of modes."""
+        return self.n_modes + self.loss_modes
+
+    @property
+    def U_full(self) -> np.ndarray:  # noqa: N802
+        """Full unitary matrix."""
+        return self._unitary
+
+    @property
+    def heralds(self) -> dict:
+        """Returns details of heralds on the input and output."""
+        return {
+            "input": copy(self._in_heralds),
+            "output": copy(self._out_heralds),
+        }
+
+    def add(self, spec: Component) -> None:
+        """Adds elements to the spec."""
+        if isinstance(spec, Loss):
+            self._loss_modes += 1
+            self._unitary = np.pad(
+                self._unitary, (0, 1), "constant", constant_values=0
+            )
+            self._unitary[-1, -1] = 1 + 0j
+        self._unitary = spec.get_unitary(self.total_modes) @ self._unitary
+
+    def add_herald(
+        self, n_photons: int, input_mode: int, output_mode: int | None = None
+    ) -> None:
+        """
+        Add a herald across a selected input/output of the circuit. If only one
+        mode is specified then this will be used for both the input and output.
+
+        Args:
+
+            n_photons (int) : The number of photons to use for the heralding.
+
+            input_mode (int) : The input mode to use for the herald.
+
+            output_mode (int | None, optional) : The output mode for the
+                herald, if this is not specified it will be set to the value of
+                the input mode.
+
+        """
+        if not isinstance(n_photons, int) or isinstance(n_photons, bool):
+            raise TypeError(
+                "Number of photons for herald should be an integer."
+            )
+        n_photons = int(n_photons)
+        if output_mode is None:
+            output_mode = input_mode
+        self._mode_in_range(input_mode)
+        self._mode_in_range(output_mode)
+        # Check if herald already used on input or output
+        if input_mode in self._in_heralds:
+            raise ValueError("Heralding already set for chosen input mode.")
+        if output_mode in self._out_heralds:
+            raise ValueError("Heralding already set for chosen output mode.")
+        # If not then update dictionaries
+        self._in_heralds[input_mode] = n_photons
+        self._out_heralds[output_mode] = n_photons
+
+    def _mode_in_range(self, mode: int) -> bool:
+        """
+        Check a mode exists within the created circuit and also confirm it
+        is an integer.
+        """
+        if not isinstance(mode, int):
+            if int(mode) == mode:
+                mode = int(mode)
+            else:
+                raise TypeError("Mode number should be an integer.")
+        if 0 <= mode < self.n_modes:
+            return True
+        raise ModeRangeError(
+            "Selected mode(s) is not within the range of the created "
+            "circuit."
+        )
