+++ lightworks/sdk/circuit/components.py
@@ -12,19 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
+from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass, fields
 
 import numpy as np
 
-from ..utils import check_unitary
+from ..utils import check_unitary, permutation_mat_from_swaps_dict
 from .parameters import Parameter
 
-# ruff: noqa: ANN202, ANN204, D101
+# ruff: noqa: ANN202, ANN204, D101, D102
 
 
-@dataclass(slots=True)
-class Component:
+@dataclass(slots=False)
+class Component(metaclass=ABCMeta):
+    @abstractmethod
+    def get_unitary(self, n_modes: int) -> np.ndarray:
+        """Returns unitary matrix for component of size n_modes."""
+
     def fields(self) -> list:
         """Returns a list of all field from the component dataclass."""
         return [f.name for f in fields(self)]
@@ -42,6 +46,9 @@ class BeamSplitter(Component):
     convention: str
 
     def __post_init__(self) -> None:
+        self.validate()
+
+    def validate(self) -> None:
         # Validate reflectivity
         if not isinstance(self.reflectivity, Parameter):
             if not 0 <= self.reflectivity <= 1:
@@ -61,6 +68,22 @@ class BeamSplitter(Component):
             return self.reflectivity.get()
         return self.reflectivity
 
+    def get_unitary(self, n_modes: int) -> np.ndarray:
+        self.validate()
+        theta = np.arccos(self._reflectivity**0.5)
+        unitary = np.identity(n_modes, dtype=complex)
+        if self.convention == "Rx":
+            unitary[self.mode_1, self.mode_1] = np.cos(theta)
+            unitary[self.mode_1, self.mode_2] = 1j * np.sin(theta)
+            unitary[self.mode_2, self.mode_1] = 1j * np.sin(theta)
+            unitary[self.mode_2, self.mode_2] = np.cos(theta)
+        elif self.convention == "H":
+            unitary[self.mode_1, self.mode_1] = np.cos(theta)
+            unitary[self.mode_1, self.mode_2] = np.sin(theta)
+            unitary[self.mode_2, self.mode_1] = np.sin(theta)
+            unitary[self.mode_2, self.mode_2] = -np.cos(theta)
+        return unitary
+
 
 @dataclass(slots=True)
 class PhaseShifter(Component):
@@ -73,11 +96,20 @@ class PhaseShifter(Component):
             return self.phi.get()
         return self.phi
 
+    def get_unitary(self, n_modes: int) -> np.ndarray:
+        unitary = np.identity(n_modes, dtype=complex)
+        unitary[self.mode, self.mode] = np.exp(1j * self._phi)
+        return unitary
+
 
 @dataclass(slots=True)
 class Loss(Component):
     mode: int
     loss: float | Parameter
+
+    def validate(self) -> None:
+        if self._loss < 0:
+            raise ValueError("Provided loss values should be greater than 0.")
 
     @property
     def _loss(self) -> float:
@@ -85,10 +117,24 @@ class Loss(Component):
             return self.loss.get()
         return self.loss
 
+    def get_unitary(self, n_modes: int) -> np.ndarray:
+        self.validate()
+        transmission = 10 ** (-self._loss / 10)
+        # Assumes loss mode to use is last mode in circuit
+        unitary = np.identity(n_modes, dtype=complex)
+        unitary[self.mode, self.mode] = transmission**0.5
+        unitary[n_modes - 1, n_modes - 1] = transmission**0.5
+        unitary[self.mode, n_modes - 1] = (1 - transmission) ** 0.5
+        unitary[n_modes - 1, self.mode] = (1 - transmission) ** 0.5
+        return unitary
+
 
 @dataclass(slots=True)
 class Barrier(Component):
     modes: list
+
+    def get_unitary(self, n_modes: int) -> np.ndarray:
+        return np.identity(n_modes, dtype=complex)
 
 
 @dataclass(slots=True)
@@ -105,6 +151,9 @@ class ModeSwaps(Component):
                 "contain the same modes."
             )
 
+    def get_unitary(self, n_modes: int) -> np.ndarray:
+        return permutation_mat_from_swaps_dict(self.swaps, n_modes)
+
 
 @dataclass(slots=True)
 class Group(Component):
@@ -113,6 +162,9 @@ class Group(Component):
     mode_1: int
     mode_2: int
     heralds: dict
+
+    def get_unitary(self, n_modes: int) -> None:  # noqa: ARG002
+        return None
 
 
 @dataclass(slots=True)
@@ -134,3 +186,11 @@ class UnitaryMatrix(Component):
         # Also check label is a string
         if not isinstance(self.label, str):
             raise TypeError("Label for unitary should be a string.")
+
+    def get_unitary(self, n_modes: int) -> np.ndarray:
+        unitary = np.identity(n_modes, dtype=complex)
+        nm = self.unitary.shape[0]
+        unitary[self.mode : self.mode + nm, self.mode : self.mode + nm] = (
+            self.unitary[:, :]
+        )
+        return unitary
